% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnoColor.R
\name{AnnoColor}
\alias{AnnoColor}
\title{AnnoColor}
\usage{
AnnoColor(type = "Nat50", num = 8)
}
\arguments{
\item{\code{num}}{how much number of colors do you want.You can provide a number or a vector. A number means you can get the first color to the num color. A vector means you only get the colors in the vector.It has no use for Heatmap Color.Default is 8.}

\item{\code{type}}{which palette do you want: "Nat50" "LNat50" "RWY" "RWB" "RYB",default is "Nat50".}
}
\value{
color code vector
}
\description{
Provide Annoroad color palette for drawing pictures.
}
\details{
5 palettes:2 for normal and 3 for heatmap

Nat50 : from nature default color

LNat50 :  Nat50 with 0.4 transparency

RWY :  Heatmap color ("red","white","yellow")(256)

RWB :  Heatmap color ("red","white","blue")(256)

RYB :  Heatmap color ("red","yellow","blue")(256)
}
\examples{
AnnoColor()
}
\author{
Ren Xue

xueren@genome.cn
}
