% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnoViewer.R
\name{AnnoViewer}
\alias{AnnoViewer}
\title{AnnoViewer}
\usage{
AnnoViewer(colorlist = AnnoColor())
}
\arguments{
\item{\code{colorlist}}{a color vector such as the result of AnnoColor()}
}
\description{
show Annoroad color palette
}
\details{
you need to provide a color vector such as the result of AnnoColor()

Default is AnnoViewer(AnnoColor())
}
\examples{
AnnoColor()
}
\author{
Ren Xue

xueren@genome.cn
}
